third_party/blink/web_tests/wpt_internal/prerender/restriction-idle-detection.https.html

<!DOCTYPE html>
<title>Prerendering cannot invoke the Idle Detection API</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  // Start prerendering a page that attempts to invoke the Idle Detection API.
  // This API is activation-gated so it's expected to fail:
  // https://wicg.github.io/nav-speculation/prerendering.html#activation-gated
  startPrerendering(`resources/idle-detection.https.html`);
  const result = await gotMessage;
  assert_equals(result, 'NotAllowedError');
  bc.close();
}, `prerendering pages should not be able to invoke the Idle Detection API`);

</script>
</body>
